+++ Source/Map/Map.hpp
@@ -72,7 +72,7 @@ public:
 	eTileTypes getTileType() const { return mParams.mTileType; }
 	eTileSub getTileSub() const { return mParams.mTileSub; }
 
-	int32 getArea() const { return mParams.mWidth * mParams.mHeight;  }
+	int32 getArea() const { return static_cast<int32>(mParams.mWidth * mParams.mHeight);  }
 	int32 getWidth() const { return static_cast<int32>(mParams.mWidth); }
 	int32 getHeight() const { return static_cast<int32>(mParams.mHeight); }
